+++ src/window/editor/build_menu.c
@@ -55,16 +55,15 @@ static const int MENU_TYPES[MENU_NUM_ITEMS][MAX_ITEMS_PER_MENU] = {
 
 static struct {
     int selected_submenu;
-    int num_items;
+    unsigned int num_items;
     int y_offset;
-
-    int focus_button_id;
+    unsigned int focus_button_id;
 } data = {MENU_NONE};
 
 static int count_items(int submenu)
 {
     int count = 0;
-    for (int i = 0; i < MAX_ITEMS_PER_MENU && MENU_TYPES[submenu][i] >= 0; i++) {
+    for (unsigned int i = 0; i < MAX_ITEMS_PER_MENU && MENU_TYPES[submenu][i] >= 0; i++) {
         count++;
     }
     return count;
@@ -92,7 +91,7 @@ static int get_sidebar_x_offset(void)
 static void draw_menu_buttons(void)
 {
     int x_offset = get_sidebar_x_offset();
-    for (int i = 0; i < data.num_items; i++) {
+    for (unsigned int i = 0; i < data.num_items; i++) {
         label_draw(x_offset - MENU_X_OFFSET, data.y_offset + MENU_Y_OFFSET + MENU_ITEM_HEIGHT * i, 10,
             data.focus_button_id == i + 1 ? 1 : 2);
         lang_text_draw_centered(48, MENU_TYPES[data.selected_submenu][i], x_offset - MENU_X_OFFSET,
@@ -113,7 +112,7 @@ static int click_outside_menu(const mouse *m, int x_offset)
         (m->x < x_offset - MENU_X_OFFSET - MENU_CLICK_MARGIN ||
             m->x > x_offset + MENU_CLICK_MARGIN ||
             m->y < data.y_offset + MENU_Y_OFFSET - MENU_CLICK_MARGIN ||
-            m->y > data.y_offset + MENU_Y_OFFSET + MENU_CLICK_MARGIN + MENU_ITEM_HEIGHT * data.num_items);
+            m->y > data.y_offset + MENU_Y_OFFSET + MENU_CLICK_MARGIN + MENU_ITEM_HEIGHT * (int) data.num_items);
 }
 
 static int handle_build_submenu(const mouse *m)
